requirements.txt
fastapi
uvicorn
celery[redis]
aio_pika
python-dotenv